refactor(webhook-runtime): move Meta hub.challenge verify into @mulmobridge/webhook-runtime (graduate to 1.0.0)
Phase 3 PR2 (stacked on #2272). The Meta webhook verify helper —
byte-identical SAFE_CHALLENGE_RE + narrowChallenge in messenger/src/verify.ts
and whatsapp/src/verify.ts — moves into the shared @mulmobridge/webhook-runtime.

- webhook-runtime: add SAFE_CHALLENGE_RE + narrowChallenge (the CodeQL
  reflected-xss sanitiser for hub.challenge, #1328).
- Delete both bridge verify.ts; messenger/whatsapp import from webhook-runtime.
- Consolidate the two near-duplicate root regression tests into
  webhook-runtime's own comprehensive suite (runs via yarn workspaces run test).
- webhook-runtime graduates 0.1.0 -> 1.0.0 (the "packages updated going forward
  start at 1.0.0+" directive, applied to @mulmobridge/* too); 6 consumer ranges
  swept ^0.1.0 -> ^1.0.0. Ungated (webhook-runtime is not a launcher dep).

relay/src/webhooks/meta.ts left as-is — Cloudflare Workers runtime can't import
the Express-based webhook-runtime.

format / lint (0 new) / typecheck / build / webhook-runtime tests 24 / common
tests 11 / deps / drift / smoke / launcher-sync all green.

Release prerequisite: publish @mulmobridge/webhook-runtime@1.0.0 before any
consumer bridge is next published.
