+++ packages/bridges/messenger/package.json
@@ -23,7 +23,7 @@
   "dependencies": {
     "@mulmobridge/client": "^0.1.5",
     "@mulmobridge/protocol": "^0.1.4",
-    "@mulmobridge/webhook-runtime": "^0.1.0",
+    "@mulmobridge/webhook-runtime": "^1.0.0",
     "@mulmoclaude/common": "^1.0.0",
     "dotenv": "^17.0.0",
     "express": "^5.1.0",
