+++ packages/webhook-runtime/test/test_webhook-runtime.ts
@@ -2,7 +2,7 @@ import { describe, it } from "node:test";
 import assert from "node:assert/strict";
 import crypto from "crypto";
 import express from "express";
-import { configureTrustProxy, createWebhookApp, createWebhookRateLimit, verifyHmacSignature } from "../src/index.ts";
+import { configureTrustProxy, createWebhookApp, createWebhookRateLimit, narrowChallenge, SAFE_CHALLENGE_RE, verifyHmacSignature } from "../src/index.ts";
 
 // Generated (not a literal) so the secret-scanner lint doesn't flag a
 // test fixture as a real credential.
@@ -84,5 +84,73 @@ describe("factory shapes", () => {
   it("createWebhookApp returns an app with x-powered-by disabled", () => {
     const app = createWebhookApp();
     assert.equal(app.get("x-powered-by"), false);
+  });
+});
+
+describe("narrowChallenge — accepted forms (Meta hub.challenge)", () => {
+  it("typical Meta nonce (alphanumeric)", () => {
+    assert.equal(narrowChallenge("ABC123xyz789"), "ABC123xyz789");
+  });
+
+  it("base64url shape (alphanumeric + _ + -)", () => {
+    assert.equal(narrowChallenge("aB-cD_eF1"), "aB-cD_eF1");
+  });
+
+  it("single char (minimum length)", () => {
+    assert.equal(narrowChallenge("a"), "a");
+  });
+
+  it("256-char string (current upper bound)", () => {
+    const long = "a".repeat(256);
+    assert.equal(narrowChallenge(long), long);
+  });
+});
+
+describe("narrowChallenge — rejected forms", () => {
+  it("rejects empty string", () => {
+    assert.equal(narrowChallenge(""), null);
+  });
+
+  it("rejects beyond 256 chars (length cap)", () => {
+    assert.equal(narrowChallenge("a".repeat(257)), null);
+  });
+
+  it("rejects non-string types (number, undefined, null)", () => {
+    assert.equal(narrowChallenge(123), null);
+    assert.equal(narrowChallenge(undefined), null);
+    assert.equal(narrowChallenge(null), null);
+  });
+
+  it("rejects array (defeats ?hub.challenge[]=... bypass)", () => {
+    // Express parses `?hub.challenge=a&hub.challenge=b` as an array;
+    // the string check at the top of narrowChallenge catches it
+    // instead of toString-coercing.
+    assert.equal(narrowChallenge(["abc"]), null);
+  });
+
+  it("rejects HTML-meta / XSS probe payloads", () => {
+    assert.equal(narrowChallenge("<script>alert(1)</script>"), null);
+    assert.equal(narrowChallenge('" onload="evil"'), null);
+    assert.equal(narrowChallenge("javascript:alert(1)"), null);
+  });
+
+  it("rejects characters outside the base64url alphabet", () => {
+    assert.equal(narrowChallenge("abc="), null);
+    assert.equal(narrowChallenge("a+b"), null);
+    assert.equal(narrowChallenge("a/b"), null);
+    assert.equal(narrowChallenge("with space"), null);
+    assert.equal(narrowChallenge("with\nnewline"), null);
+  });
+
+  it("rejects non-ASCII characters", () => {
+    assert.equal(narrowChallenge("café"), null);
+    assert.equal(narrowChallenge("日本語"), null);
+  });
+});
+
+describe("SAFE_CHALLENGE_RE — anchor sanity", () => {
+  it("regex is fully anchored (no partial-match leakage)", () => {
+    assert.equal(SAFE_CHALLENGE_RE.source.startsWith("^"), true);
+    assert.equal(SAFE_CHALLENGE_RE.source.endsWith("$"), true);
   });
 });
